+++ R/simulate_reproductivenumber.R
@@ -36,7 +36,7 @@ reproductivenumberode <- function(t, y, parms)
 #' @param e efficacy of vaccine, given as fraction between 0 and 1
 #' @param b level/rate of infectiousness for hosts in the I compartment
 #' @param g rate at which a person leaves the I compartment
-#' @param w rate at which recovered persons loose immunity and return to susceptible state
+#' @param w rate at which recovered persons lose immunity and return to susceptible state
 #' @param m the rate at which new individuals enter the model (are born)
 #' @param n the rate of natural death (the inverse it the average lifespan)
 #' @param tmax maximum simulation time, units depend on choice of units for your
@@ -52,13 +52,20 @@ reproductivenumberode <- function(t, y, parms)
 #'   something nonsensical (e.g. negative values or fractions > 1),
 #'   the code will likely abort with an error message.
 #' @examples
-#'   # To run the simulation with default parameters just call this function
+#'   # To run the simulation with default parameters just call the function:
 #'   result <- simulate_reproductivenumber()
-#'   # To choose parameter values other than the standard one, specify them e.g. like such
+#'   # To choose parameter values other than the standard one, specify them like such:
 #'   result <- simulate_reproductivenumber(S0 = 2000, I0 = 10, tmax = 100, g = 0.5, n = 0.1)
-#'   # You should then use the simulation result returned from the function, e.g. like this:
-#'   plot(result[,1],result[,2],xlab='Time',ylab='Number Susceptible',type='l')
-#' @seealso The UI of the shiny app 'ReproductiveNumber', which is part of this package, contains more details on the model.
+#'   # You should then use the simulation result returned from the function, like this:
+#'   plot(result$ts[ , "Time"],result$ts[ , "S"],xlab='Time',ylab='Number Susceptible',type='l')
+#'   # We might also want to have infectiousness rate for hosts in the I compartment
+#'   # of 0.4.
+#'   result <- simulate_reproductivenumber(S0 = 2000, I0 = 10, tmax = 100, b = 0.4)
+#'   plot(result$ts[,"Time"],result$ts[,"S"],xlab = "Time", ylab = "Number Susceptible",type="l")
+#'   # We could also have infectiousness rate of 0.6 and recovery rate of 0.2.
+#'   result <- simulate_reproductivenumber(S0 = 2000, I0 = 10, tmax = 100, b = 0.6, g = 0.2)
+#'   plot(result$ts[,"Time"],result$ts[,"S"],xlab = "Time", ylab = "Number Susceptible",type="l")
+#' @seealso The UI of the Shiny app 'ReproductiveNumber', which is part of this package, contains more details on the model.
 #' @references See e.g. Keeling and Rohani 2008 for SIR models and the
 #'   documentation for the deSolve package for details on ODE solvers
 #' @author Andreas Handel
@@ -68,7 +75,7 @@ reproductivenumberode <- function(t, y, parms)
 simulate_reproductivenumber <- function(S0 = 1000, I0 = 1, f = 0.0, e = 0.0, tmax = 300, g = 10, b = 1e-2, m = 0, n = 0, w = 0){
 
   S0eff = (1 - f*e) * S0;
-  R0 = f*e * S0;
+  R0 = f*e * S0; #initial number of recovered/removed (inlcudes vaccinated)
   Y0 = c(S = S0eff, I = I0, R = R0);  #combine initial conditions into a vector
   dt = min(0.1, tmax / 1000); #time step for which to get results back
   timevec = seq(0, tmax, dt); #vector of times for which solution is returned (not that internal timestep of the integrator is different)
@@ -80,6 +87,10 @@ simulate_reproductivenumber <- function(S0 = 1000, I0 = 1, f = 0.0, e = 0.0, tma
   #the result is saved in the odeoutput matrix, with the 1st column the time, the 2nd, 3rd, 4th column the variables S, I, R
   #This odeoutput matrix will be re-created every time you run the code, so any previous results will be overwritten
   odeoutput = deSolve::ode(y = Y0, times = timevec, func = reproductivenumberode, parms=pars, method = "vode", atol=1e-8, rtol=1e-8);
+  
+  colnames(odeoutput) <- c('Time','S','I','R') 
+  result <- list()
+  result$ts <- as.data.frame(odeoutput)
 
-  return (odeoutput)
+  return(result)
 }