+++ NEWS.md
@@ -1,12 +1,41 @@
-# DSAIDE version 0.6.0
-* Various bug fixes in the apps and documentation
-* Updated "what to do" instructions for multiple apps
+# DSAIDE 0.7.0
+
+## User facing changes
+
+* Major changes to underlying plotting and text generating approaches, now uses ggplot, shows better plots and more informative text.
+
+* Added a function that lets one call apps without going through the main menu.
+
+* Added new apps.
+
+* Updated a lot of the instructions and documentation.
+
+* Fixed bug that prevented opening of apps from the main menu for some versions of shiny.
+
+## Other changes
+
+* Standardized all simulator functions to return lists.
+
+* Simplified underlying code and documentation structure of shiny apps.
+
+* Added some unit tests.
+
+* Lots of other code cleanup to make things more robust and streamlined
 
 
-# DSAIDE version 0.5.0
-* Added one new app
-* Various bug fixes in the apps and documentation
+# DSAIDE 0.6.0
+
+* Various bug fixes in the apps and documentation.
+
+* Updated "what to do" instructions for multiple apps.
+
+# DSAIDE 0.5.0
+
+* Added one new app.
+
+* Various bug fixes in the apps and documentation.
 
 
-# DSAIDE version 0.4.0
-* This is the first release to CRAN
+# DSAIDE 0.4.0
+
+* This is the first release to CRAN.